Reorganize test files to one per tested class
